+++ src/Contracts/Translation.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Khamsolt\Orchid\Files\Contracts;
+
+interface Translation
+{
+    public function get(string $text, array $replace = [], string $locale = null): string;
+}