+++ src/Http/Requests/UpdateRequest.php
@@ -4,7 +4,7 @@ namespace Khamsolt\Orchid\Files\Http\Requests;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class UpdateRequest extends FormRequest
+final class UpdateRequest extends FormRequest
 {
     public function authorize(): bool
     {