+++ src/Providers/AuthServiceProvider.php
@@ -3,19 +3,15 @@
 namespace Khamsolt\Orchid\Files\Providers;
 
 use Illuminate\Support\ServiceProvider;
-use Khamsolt\Orchid\Files\Contracts\Permissions;
+use Khamsolt\Orchid\Files\Contracts\Authorization;
 use Orchid\Platform\Dashboard;
 
 class AuthServiceProvider extends ServiceProvider
 {
-    protected Dashboard $dashboard;
-
-    public function boot(Dashboard $dashboard, Permissions $permissions): void
+    public function boot(Dashboard $dashboard, Authorization $authorization): void
     {
-        $this->dashboard = $dashboard;
-
-        $this->app->booted(function () use ($permissions) {
-            $this->dashboard->registerPermissions($permissions->getItemPermission());
+        $this->app->booted(function () use ($dashboard, $authorization) {
+            $dashboard->registerPermissions($authorization->getItemPermission());
         });
     }
 }