+++ src/Screens/FileViewScreen.php
@@ -2,11 +2,12 @@
 
 namespace Khamsolt\Orchid\Files\Screens;
 
-use Illuminate\Contracts\Config\Repository;
-use Illuminate\Contracts\Translation\Translator;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Routing\Redirector;
-use Khamsolt\Orchid\Files\Contracts\Permissions;
+use Khamsolt\Orchid\Files\Contracts\Authorization;
+use Khamsolt\Orchid\Files\Contracts\Configuration;
+use Khamsolt\Orchid\Files\Contracts\Translation;
+use Khamsolt\Orchid\Files\Enums\Action;
 use Khamsolt\Orchid\Files\FileSettings;
 use Khamsolt\Orchid\Files\Models\Attachment;
 use Khamsolt\Orchid\Files\View\Components\Preview;
@@ -18,21 +19,22 @@ use Orchid\Screen\Layouts\Persona;
 use Orchid\Screen\Screen;
 use Orchid\Screen\Sight;
 use Orchid\Support\Presenter;
+use Throwable;
 
-class FileViewScreen extends Screen
+final class FileViewScreen extends Screen
 {
-    public int $id;
+    public int|null $id = null;
 
-    public string $url;
+    public string|null $url = null;
 
     public bool $isImage = false;
 
     public function __construct(
-        private readonly Repository  $config,
-        private readonly Permissions $permissible,
-        private readonly Redirector  $redirector,
-        private readonly Translator  $translator,
-        private readonly Toast       $toast
+        private readonly Authorization $authorization,
+        private readonly Configuration $configuration,
+        private readonly Translation $translator,
+        private readonly Redirector $redirector,
+        private readonly Toast $toast
     ) {
     }
 
@@ -48,7 +50,7 @@ class FileViewScreen extends Screen
 
     public function permission(): ?iterable
     {
-        return $this->permissible->accessViewFile();
+        return $this->authorization->authorize(Action::VIEW);
     }
 
     public function query(Attachment $attachment): array
@@ -66,27 +68,21 @@ class FileViewScreen extends Screen
 
     public function commandBar(): array
     {
-        /** @var string $deleteConfirmMessage */
-        $deleteConfirmMessage = $this->translator->get('Once the media file is deleted, all of its resources and data will be permanently deleted. Before deleting your media file, please download any data or information that you wish to retain.');
-
-        /** @var string $urlView */
-        $urlView = $this->config->get('orchid-files.routes.edit');
-
         return [
             Button::make('Delete')
-                ->confirm($deleteConfirmMessage)
-                ->icon('trash')
+                ->confirm($this->translator->get('Once the media file is deleted, all of its resources and data will be permanently deleted. Before deleting your media file, please download any data or information that you wish to retain.'))
+                ->icon('bs.file-minus-fill')
                 ->method('delete'),
 
             Link::make('Open')
                 ->target('blank')
-                ->icon('cloud-download')
+                ->icon('bs.file-arrow-down-fill')
                 ->href($this->url),
 
             Link::make('Edit')
                 ->target('blank')
-                ->icon('pencil')
-                ->route($urlView, [$this->id]),
+                ->icon('bs.file-break-fill')
+                ->route($this->configuration->route(Action::EDIT), [$this->id]),
         ];
     }
 
@@ -97,77 +93,59 @@ class FileViewScreen extends Screen
                 ->canSee($this->isImage),
 
             LayoutFactory::legend('attachment', [
-                Sight::make('id', $this->translate('#ID')),
+                Sight::make('id', $this->translator->get('#ID')),
 
-                Sight::make('user_id', $this->translate('User'))
+                Sight::make('user_id', $this->translator->get('User'))
                     ->render(function (Attachment $attachment) {
                         if ($attachment->getRelation('user') === null) {
                             return null;
                         }
-                        /** @var array<string, class-string> $presenters */
-                        $presenters = $this->config->get('orchid-files.presenters');
 
-                        $filePresenter = new FileSettings();
+                        $filePresenter = new FileSettings($this->configuration);
 
-                        $presenter = $filePresenter->resolveUserPresenter($attachment, $presenters);
+                        $presenter = $filePresenter->resolveUserPresenter($attachment);
 
                         if ($presenter instanceof Presenter) {
                             return new Persona($presenter);
                         }
 
-                        return $this->translate('User ID :id', ['id' => $presenter]);
+                        return $this->translator->get('User ID :id', ['id' => $presenter]);
                     }),
 
-                Sight::make('name', $this->translate('Name')),
-                Sight::make('original_name', $this->translate('Title')),
-                Sight::make('mime', $this->translate('Mime')),
-                Sight::make('extension', $this->translate('Extension')),
+                Sight::make('name', $this->translator->get('Name')),
+                Sight::make('original_name', $this->translator->get('Title')),
+                Sight::make('mime', $this->translator->get('Mime')),
+                Sight::make('extension', $this->translator->get('Extension')),
 
-                Sight::make('size', $this->translate('Size'))
-                    ->render(fn (Attachment $attachment) => $attachment->sizeToKb() . ' Kb'),
+                Sight::make('size', $this->translator->get('Size'))
+                    ->render(fn(Attachment $attachment) => $attachment->sizeToKb().' Kb'),
 
-                Sight::make('sort', $this->translate('Sort')),
-                Sight::make('path', $this->translate('Path')),
-                Sight::make('description', $this->translate('Description')),
-                Sight::make('alt', $this->translate('Alt')),
-                Sight::make('hash', $this->translate('Hash')),
-                Sight::make('disk', $this->translate('Disk')),
-                Sight::make('group', $this->translate('Group')),
+                Sight::make('sort', $this->translator->get('Sort')),
+                Sight::make('path', $this->translator->get('Path')),
+                Sight::make('description', $this->translator->get('Description')),
+                Sight::make('alt', $this->translator->get('Alt')),
+                Sight::make('hash', $this->translator->get('Hash')),
+                Sight::make('disk', $this->translator->get('Disk')),
+                Sight::make('group', $this->translator->get('Group')),
 
-                Sight::make('created_at', $this->translate('Created'))
-                    ->render(fn (Attachment $attachment) => $attachment->created_at->toDateTimeString()),
+                Sight::make('created_at', $this->translator->get('Created'))
+                    ->render(fn(Attachment $attachment) => $attachment->created_at->toDateTimeString()),
 
-                Sight::make('updated_at', $this->translate('Updated'))
-                    ->render(fn (Attachment $attachment) => $attachment->updated_at->toDateTimeString()),
+                Sight::make('updated_at', $this->translator->get('Updated'))
+                    ->render(fn(Attachment $attachment) => $attachment->updated_at->toDateTimeString()),
             ]),
         ];
     }
 
+    /**
+     * @throws Throwable
+     */
     public function delete(Attachment $attachment): RedirectResponse
     {
         $attachment->deleteOrFail();
 
-        /** @var string $successMessage */
-        $successMessage = $this->translator->get('Media file deleted successfully');
+        $this->toast->success($this->translator->get('Media file deleted successfully'));
 
-        $this->toast->success($successMessage);
-
-        /** @var string $redirectListUrl */
-        $redirectListUrl = $this->config->get('orchid-files.routes.list');
-
-        return $this->redirector->route($redirectListUrl);
-    }
-
-    /**
-     * @param string $text
-     * @param array<string, string|int|float> $replace
-     * @return string
-     */
-    private function translate(string $text, array $replace = []): string
-    {
-        /** @var string $translated */
-        $translated = $this->translator->get($text, $replace);
-
-        return $translated;
+        return $this->redirector->route($this->configuration->route(Action::LIST));
     }
 }